+++ libhesai/PtcParser/include/ptc_2_0_parser.h
@@ -106,7 +106,7 @@ public:
 
   uint8_t GetHeaderIdentifier0() { return header_.identifier0_; }
   uint8_t GetHeaderIdentifier1() { return header_.identifier1_; }
-  int GetHeaderSize() { return sizeof(PTCHeader_2_0); }
+  uint16_t GetHeaderSize() { return sizeof(PTCHeader_2_0); }
   uint8_t GetHeaderReturnCode() const { return header_.return_code_; }
   uint8_t GetHeaderCmd() const {return header_.cmd_; }
   uint32_t GetHeaderPayloadLen() const { return header_.GetPayloadLen();}
