+++ libhesai/UdpParserGpu/src/udp3_1_parser_gpu.cu
@@ -187,7 +187,7 @@ int Udp3_1ParserGpu<T_Point>::LoadCorrectionFile(std::string lidar_correction_fi
   if (fin.is_open()) {
     int length = 0;
     fin.seekg(0, std::ios::end);
-    length = fin.tellg();
+    length = static_cast<int>(fin.tellg());
     fin.seekg(0, std::ios::beg);
     char *buffer = new char[length];
     fin.read(buffer, length);
